Make various fixes to falconctl parsing to remove errors on selecting from the falconctl_options table
Fixes #32239.

This changes tags to return a comma-delimited list on multiple tags, the single tag when there's only one, and "is not set" (similar to other values) when no tags are set.

Confirmed that this allows us to run `SELECT * FROM falconctl_options` without issue on various configurations of Crowdstrike Falcon on Linux.