s/TeamWithoutExtras/TeamLite
For grepping, it makes more sense to have something that doesn't conflict with a lot of other code, so `Team` isn't coming back.

This also gets us a step toward having a TeamLite data structure that we can hint against, avoiding unpleasant surprises when fields we thought got populated on a Team are actually zero values.