Add Ubuntu Advantage Query to Library (#35680)
Adds the Ubuntu Advantage policy to the policy library that checks to see
if the file exists, and that it is attached, and expiry date has not
passed.

This can be used in combination with a script automation running `pro
attach <tokenID>`, with a tokenID configured in Fleet Variables. E.g:
`pro attach $FLEET_SECRET_UBUNTUPRO` for remediation.

---------

Co-authored-by: Noah Talerman <47070608+noahtalerman@users.noreply.github.com>