Add test to validate mock changes (#35663)
See
https://github.com/fleetdm/fleet/actions/runs/19313106183/job/55238111566?pr=35663
for an example of what this looks like when it fails

See https://github.com/fleetdm/fleet/pull/35655 for why we need this.
Someone was able to check in code that inadvertently broke things for
other developers. This is similar to why we have the existing "DB" tests
that validate you updated the DB Schema file