Fix table of content links on Rest API page (#35348)
Fixes the outdated MDM link, and adds the missing top-level links.

_The notation for integrations `#integrations-1`, is to take the second
occurence of integrations, as we have another integrations title in the
update webhook configuration._

I can see this doesn't affect the website (or shouldn't) since it uses
it's own way to generate the TOC, but will help for manually browsing
the markdown.