Improve mobile web login experience: Update LoginForm.tsx (#36172)
I don't know if this works (didn't check the component interface to see
what props it takes, so will leave that to up to the reviewer of this PR
to edit and complete, if you please wouldn't mind. Also if you're aware
of other areas we can make this improvement to the email field, such as
signup and in the password recovery flow, then would you please make
those changes?)

Goal: I'm trying to improve the mobile web experience and get handset
devices like iPhones to open the email keyboard and _not_ capitalize the
first letter you type when logging in.

To do that, we need the generated HTML to use `type="email"`. For
example, to look like this:
https://github.com/fleetdm/fleet/blob/78ee32fd588ab273fa2cc890487b614be8c5c754/website/views/pages/entrance/login.ejs#L19


- [x] QA'd all new/changed functionality manually

---------

Co-authored-by: Jacob Shandling <jacob@shandling.dev>