Changing "macOS Setup Assistant" to "Setup Assistant" in frontend and guide (#36296)
We recently implemented this change:
https://github.com/fleetdm/fleet/pull/35782

I think it makes sense to capitalize Setup Assistant because it is a
proper noun. However, we should drop macOS from the name since the Setup
Assistant exists on iOS/iPadOS too and the skip panes functionality is
platform agnostic. Otherwise users may be confused on how they achieve
this on their iOS/iPadOS devices since this is so explicitly labeled as
macOS.