Use writer for reads for software icon activity population (#36552)
Fixes #36090. Probably overkill to read both queries from the writer but
given that this isn't a hot path it doesn't really hurt, and ensures
consistency even in higher replica-lag situations where a software title
is added/updated immediately before the associated icon is uploaded.

# Checklist for submitter

If some of the following don't apply, delete the relevant line.

## Testing

- [x] QA'd all new/changed functionality manually

For unreleased bug fixes in a release candidate, one of:

- [x] Confirmed that the fix is not expected to adversely impact load
test results