Update how we version Google Chrome for macOS (#36458)
- Updating how we version Google Chrome for macOS. The current process
of using "latest" as the version number means we are never validating
the pkg we are installing. It also doesn't display a good pkg version
for admins to track. This change may introduce a small window where the
pkg that gets installed is newer than the version we represent on our
website or in our library. This is an accepted risk since we can say we
have validated up to that version.