Freeze versioning on Adobe Acrobat Pro (macOS) (#36609)
Freezing Adobe Acrobat Pro on macOS. They recently updated the product
but did not update the installer leading to validation issues on our
end. I am freezing this and will check back in a few days to see if it
has been updated or if we need to update our validation.