Add activities for enabling/disabling "update new hosts to latest" (#37089)
<!-- Add the related story/sub-task/bug number, like Resolves #123, or
remove if NA -->
**Related issue:** Resolves #37088

# Details

This just adds the new activities for use in the API handler and the
front-end. I neglected to include this in the spec so I'm putting up a
quick PR to keep from adding to the back-end ticket scope.

# Checklist for submitter

No checklist items apply here; it just needs to implement what's laid
out in https://github.com/fleetdm/fleet/pull/35777/files.

---------

Co-authored-by: Juan Fernandez <juan@fleetdm.com>