Update Company Portal icon matching (#37334)
This pull request enhances the icon mapping logic for software names in
the `SoftwarePage` component, specifically ensuring that the Microsoft
Company Portal app is correctly associated with its icon, even when
referenced by its package name.

**Icon mapping improvements:**

* Added `"microsoft.companyportal"` to the `SOFTWARE_NAME_TO_ICON_MAP`
so that this package name is mapped to the `IntuneCompanyPortal` icon.
* Updated the `matchStrictNameSourceToIcon` function to explicitly
return the `IntuneCompanyPortal` icon when the software name is exactly
`"microsoft.companyportal"`.