Update using-maintenance-windows.md
I removed the feature image since it adds little value and pushes the content further down the page.

Images can be optionally defined in the metadata for social sharing thumbnail, but are not required. The default thumbnail will be used if the article image is not defined in the metadata.