Update zero-trust-attestation-with-fleet.md
I removed the feature image since it adds little value and pushes the content further down the page.

Articles and guides do not require a feature image. Social sharing thumbnails can be optionally defined in the metadata, but are not required. The default thumbnail will be used if the article image is not defined in the metadata.