Add `no_check` to Google Chrome FMA for Windows (#37380)
This pull request updates the configuration for the Google Chrome
maintained app on Windows to bypass hash checking for the installer.
This change ensures that installations will proceed even if the
installer hash changes, which can help avoid issues with frequent
upstream updates.

Configuration changes for hash checking:

* Added `"ignore_hash": true` to the `winget/google-chrome.json` input
file to instruct the system to skip hash verification during
installation.
* Set `"sha256": "no_check"` in the `outputs/google-chrome/windows.json`
output file to explicitly indicate that the installer hash should not be
checked.