Allow disabling schedule with empty start/end time (#37709)
<!-- Add the related story/sub-task/bug number, like Resolves #123, or
remove if NA -->
**Related issue:** Resolves #37599

# Details

Fixes an unreleased issue where if you clear the start/end time of a
software auto-update schedule and then attempt to disable and save the
schedule, the API returns an error due to the start/end time being
invalid. Now, the start/end time will be unchanged if the schedule is
disabled.

## Testing

- [X] Added/updated automated tests
- [X] QA'd all new/changed functionality manually

For unreleased bug fixes in a release candidate, one of:

- [X] Confirmed that the fix is not expected to adversely impact load
test results
- [X] Alerted the release DRI if additional load testing is needed